fix pylonconfig setup

+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/config/PylonConfig.kt
@@ -13,5 +13,5 @@ object PylonConfig {
     val allowedBlockErrors = config.get("tick-delay", 10)
 
     @JvmField
-    val itemTickRate = config.get("item-tick-rate", 15L)
+    val itemTickRate = config.get("item-tick-rate", 15).toLong()
 }
